fix - double 중복 제거

+++ src/main/java/lotto/domain/LottoResult.java
@@ -33,7 +33,7 @@ public class LottoResult {
             total += lottoRank.getReward() * result.get(lottoRank);
         }
 
-        return (double)total / (double)money;
+        return (double)total / money;
 
     }
 }
